refactor: extract matching
